backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.3.1
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==1.26.2
pandas==2.1.3
pyarrow==22.0.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets
google-genai==0.3.0
